\name{backLog}
\alias{backLog}
\title{
Backtransforms coefficient estimates from logistic regression
}
\description{
Takes raw coefficient values (e.g. from a logistic regression) and returns back-transformed estimates.
}
\usage{
backLog(x)
}
\arguments{
  \item{x}{
A numerical value to be back-transformed.
}
}

\value{
  Returns the back-transformed estimate.
}

\author{
Kevin Arbuckle
}