\name{sample.tree}
\alias{sample.tree}
\docType{data}
\title{
Phylogenetic tree to accompany sample.data
}
\description{
A simulated phylogeny from which the sample.data dataset was simulated, for use with the Wheatsheaf index functions in the windex package.
}
\usage{data(sample.tree)}
\format{
  Phylogenetic tree of the class 'phylo' with 100 tips and (ultrametric) branch lengths.
}

\examples{
data(sample.tree)
summary(sample.tree)
plot(sample.tree)
}
\keyword{datasets}
