\name{se}
\alias{se}
\title{
Calculates standard error
}
\description{
Calculates standard error of a numerical vector.
}
\usage{
se(x)
}
\arguments{
  \item{x}{
Numerical vector.
}
}

\value{
  Returns the standard error of the values in the vector.
}

\author{
Kevin Arbuckle
}